moving all the text parsing method
